decentralized-kyc/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.2
web3==7.4.0
sqlalchemy==2.0.27
alembic==1.13.1
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.1
eth-account==0.13.0
aiohttp==3.9.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
